+++ spec/UIManager.spec.ts
@@ -13,6 +13,7 @@ import { MobileV3PlayerEvent } from '../src/ts/utils/MobileV3PlayerAPI';
 import { UIContainer } from '../src/ts/components/UIContainer';
 import { Container } from '../src/ts/components/Container';
 import { StorageUtils } from '../src/ts/utils/StorageUtils';
+import { prefixCss } from '../src/ts/components/DummyComponent';
 import { RecommendationConfig, TimelineMarker } from '../src/ts/UIConfig';
 import { FullscreenToggleButton } from '../src/ts/components/buttons/FullscreenToggleButton';
 
@@ -643,5 +644,5 @@ function resumeStorageKeyForSourceIdentifier(sourceIdentifier: string): string {
     hash = (hash << 5) - hash + sourceIdentifier.charCodeAt(i);
     hash |= 0;
   }
-  return `bitmovin.player.ui.resume.${(hash >>> 0).toString(36)}`;
+  return `${prefixCss('resume')}.${(hash >>> 0).toString(36)}`;
 }